% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_cmpas_by_time}
\alias{cimiss_cmpas_by_time}
\title{Retrieve CMPAS data file URL or download to local directory.}
\usage{
cimiss_cmpas_by_time(
  times,
  dataCode = "SURF_CMPA_RT_NC",
  outdir = NULL,
  limitCnt = NULL,
  resolution = "0P05"
)
}
\arguments{
\item{times}{: times for retrieve, 'YYYYMMDDHHMISS,YYYYMMDDHHMISS,...'}

\item{dataCode}{: dataset code, like "SURF_CMPA_RT_NC, SURF_CMPA_NRT_NC, ...'}

\item{outdir}{: local output directory.}

\item{limitCnt}{: the number of maximum returned files.}

\item{resolution}{: the CMPAS resolution, 0P05 for 0.5 degree,
0P01 for 0.1 degree, others for all.}
}
\value{

}
\description{
Retrieve CMPAS data file URL or download to local directory.
}
\examples{
     files <- cimiss_cmpas_by_time('20200831000000')
}
