% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_download_file}
\alias{cimiss_download_file}
\title{Download data file from URL}
\usage{
cimiss_download_file(url, filename, outdir = ".")
}
\arguments{
\item{url}{: file URL.}

\item{filename}{: local file name.}

\item{outdir}{: local directory.}
}
\value{

}
\description{
Download data file from URL
}
