% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_get_model_latest_time}
\alias{cimiss_get_model_latest_time}
\title{Get the model latest initial time.}
\usage{
cimiss_get_model_latest_time(
  dataCode = "NAFP_FOR_FTM_HIGH_EC_ANEA",
  latestTime = 24
)
}
\arguments{
\item{dataCode}{: dataset code, like "NAFP_FOR_FTM_HIGH_EC_ANEA", "NAFP_FOR_FTM_HIGH_EC_ASI", and so on.}

\item{latestTime}{: latestTime, >0, like 2 is return the latest data in 2 hours.}
}
\value{
data frame of observations.
}
\description{
\preformatted{Retrieve data by cimiss music REST API.
Refer to http://10.20.76.55/cimissapiweb/index_index.action
}
}
\examples{
  data <- cimiss_get_model_latest_time('NAFP_FOR_FTM_HIGH_EC_ANEA', latestTime=12)

}
