% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_model_grid_in_rect}
\alias{cimiss_model_grid_in_rect}
\title{Retrieve grid data from CIMISS service.}
\usage{
cimiss_model_grid_in_rect(
  time,
  minLat,
  minLon,
  maxLat,
  maxLon,
  dataCode = "NAFP_FOR_FTM_HIGH_EC_GLB",
  fcstLevel = 0,
  fcstEle = "TEF0",
  validTime = 0,
  outList = FALSE
)
}
\arguments{
\item{time, }{model run time, like "2016081712"}

\item{minLat, }{minimum latitude}

\item{minLon, }{minimum longitude}

\item{maxLat, }{maximum latitude}

\item{maxLon, }{maximum longitude}

\item{dataCode, }{MUSIC data code, default is "NAFP_FOR_FTM_HIGH_EC_GLB", which is for
ECMWF deterministic high resolution model global surface forecast.}

\item{fcstLevel, }{vertical level, default is 0.}

\item{fcstEle, }{forecast element, default is 2m temperature "TEF0"}

\item{validTime, }{forecast hour, default is 0}
}
\value{
list(x,y,z,level,time,fhour)
}
\description{
\preformatted{Retrieve grid data from CIMISS service with MUSIC java libray
}
}
\examples{
  data <- cimiss_model_grid_in_rect("2016081712", 15, 60, 55, 136)


}
