% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_model_point_by_time_and_level}
\alias{cimiss_model_point_by_time_and_level}
\title{retrieve model point forecast data from CIMISS.}
\usage{
cimiss_model_point_by_time_and_level(time,
  dataCode = "NAFP_FOR_FTM_HIGH_EC_GLB", fcstEle = "TEF0",
  fcstLevel = 0, minVT = 0, maxVT = 240, latLons = "39.8/116.4667")
}
\arguments{
\item{time, }{run time for retrieve, 'YYYYMMDDHHMISS,YYYYMMDDHHMISS,...'}

\item{dataCode, }{like "NAFP_FOR_FTM_HIGH_EC_GLB", "NAFP_FOR_FTM_HIGH_EC_ASI", and so on.}

\item{fcstEle, }{forecast element (single), default is 2m temperature "TEF0"}

\item{fcstLevel, }{forecast level (single)}

\item{minVT, }{start forecast hour}

\item{maxVT, }{end forecast hour}

\item{latLons, }{latitude/longitude, multiple points are supported with separated by ",".}
}
\value{
data.frame of point forecast.
}
\description{
retrieve model point forecast data from CIMISS.
}
\examples{
  dataV <- cimiss_model_point_by_time_and_level("20170120000000")

}
