% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_obs_by_time}
\alias{cimiss_obs_by_time}
\title{Retrieve station records from CIMISS by times.}
\usage{
cimiss_obs_by_time(
  times,
  dataCode = "SURF_CHN_MUL_HOR_N",
  elements = "Station_Id_C,Station_Id_d,lat,lon,Datetime,TEM"
)
}
\arguments{
\item{times}{: time for retrieve, 'YYYYMMDDHHMISS,YYYYMMDDHHMISS,...'}

\item{dataCode}{: dataset code, like "SURF_CHN_MUL_HOR", "SURF_CHN_MUL_HOR_N", and so on.}

\item{elements}{: elements for retrieve, 'ele1,ele2,...'}

\item{staIds}{: station ids, 'xxxxx,xxxxx,...'}
}
\value{
data frame of observations.
}
\description{
\preformatted{Retrieve data by cimiss music API.
You should install MUSIC java client.
1) download MUSIC java client from http://10.20.76.55/cimissapiweb/page/apipages/ClientDocumentRight.jsp
2) set environment variable
CLASSPATH=C:/soft/music-lib/Ice.jar;C:/soft/music-lib/music-client-v1.4.0.jar;C:/soft/music-lib
3) call java function
}
}
\examples{
\dontrun{
  r1 <- cimiss_obs_by_time('20181108000000')
}

}
