% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_obs_by_time_range_and_staID}
\alias{cimiss_obs_by_time_range_and_staID}
\title{Retrieve station records from CIMISS by time and station ID.}
\usage{
cimiss_obs_by_time_range_and_staID(
  timeRange,
  hourSeparate = NULL,
  minSeparate = NULL,
  dataCode = "SURF_CHN_MUL_HOR_N",
  elements = "Station_Id_C,Datetime,TEM",
  staIds = "54511"
)
}
\arguments{
\item{timeRange}{: time range for retrieve, "\link{YYYYMMDDHHMISS,YYYYMMDDHHMISS}",
like, "\link{201509010000,20150903060000}"}

\item{dataCode}{: dataset code, like "SURF_CHN_MUL_HOR", "SURF_CHN_MUL_HOR_N", and so on.}

\item{elements}{: elements for retrieve, 'ele1,ele2,...'}

\item{staIds}{: station ids, 'xxxxx,xxxxx,...'}

\item{hourSeparate:}{hour space, \link{1, 24}, like 6: 0,6,12,18}

\item{minSeparate:}{minute space, \link{1, 60}, like 10: 0,10,20,30,40,50}
}
\value{
data frame of observations.
}
\description{
\preformatted{Retrieve data by cimiss music API.
You should install MUSIC java client.
1) download MUSIC java client from http://10.20.76.55/cimissapiweb/page/apipages/ClientDocumentRight.jsp
2) set environment variable
CLASSPATH=C:/soft/music-lib/Ice.jar;C:/soft/music-lib/music-client-v1.4.0.jar;C:/soft/music-lib
3) call java function
}
}
\examples{
\dontrun{
  r1 <- cimiss_obs_by_time_range_and_staID("[20060801000000,20060801000000]")
}

}
