% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_obs_convert_type}
\alias{cimiss_obs_convert_type}
\title{Convert observation data.frame to correct data type.}
\usage{
cimiss_obs_convert_type(obsData)
}
\arguments{
\item{obsData}{: data frame of observations.}
}
\value{
data frame of observations.
}
\description{
Convert observation data.frame to correct data type.
}
