% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_obs_in_admin_by_time}
\alias{cimiss_obs_in_admin_by_time}
\title{Retrieve station records from CIMISS in adminstration region by time.}
\usage{
cimiss_obs_in_admin_by_time(
  times,
  adminCodes = "110000",
  dataCode = "SURF_CHN_MUL_HOR_N",
  staLevels = NULL,
  eleValueRanges = NULL,
  orderby = NULL,
  limitCnt = NULL,
  elements = "Station_Id_C,Datetime,Lat,Lon,TEM"
)
}
\arguments{
\item{times}{: times for retrieve, 'YYYYMMDDHHMISS,YYYYMMDDHHMISS,...', seperate by ","}

\item{dataCode}{: dataset code, like "SURF_CHN_MUL_HOR", "SURF_CHN_MUL_HOR_N", and so on.}

\item{elements}{: elements for retrieve, 'ele1,ele2,...'}

\item{adminCodes:}{administration(or province code), sperated by ",",
like "110000" is Beijing, "440000" is Guangdong}

\item{staLevels:}{station levels, seperated by ',', "011,012,013" is National Reference Climate Station,
Basic weather station, General weather station, ...}

\item{eleValueRanges:}{elements value ranges, seperated by ';'
range, (a,) is >a, [a,) is >=a, (,a) is <a, (,a] is <=a,
(a,b) is >a & <b, [a,b) is >=a & <b, (a,b] is >a & <=b,
[a,b] is >=a & <=b
list, a,b,c;
e.g., "VIS:(,1000);RHU:(70,)", "Q_PRE_1h:0,3,4" is PRE quantity is credible.

[a,) is >=a, (,a) is <a, (,a]: R:a,)\%20is\%20\%3E=a,\%20(,a)\%20is\%20\%3Ca,\%20(,a
[a,b) is >=a & <b, (a,b]: R:a,b)\%20is\%20\%3E=a\%20&\%20\%3Cb,\%20(a,b
[a,b]: R:a,b}

\item{orderby:}{elements order, seperated by ',', like
"TEM:asc" is ascending order temperature,
"TEM:asc,SUM_PRE_1h:desc" is ascending order temperature first and descending PRE_1h.}

\item{limitCnt:}{the number of maximum returned records}
}
\value{
data frame of observations.
}
\description{
\preformatted{Retrieve data by cimiss music REST API.
Refer to http://10.20.76.55/cimissapiweb/index_index.action
}
}
\examples{
  data <- cimiss_obs_in_admin_by_time('20200206000000', adminCodes="110000")

}
