% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_obs_in_Rect_by_time_range}
\alias{cimiss_obs_in_Rect_by_time_range}
\title{Retrieve station records from CIMISS by time and station ID.}
\usage{
cimiss_obs_in_Rect_by_time_range(timeRange, minLat, minLon, maxLat, maxLon,
  dataCode = "SURF_CHN_MUL_HOR_N",
  elements = "Station_Id_C,Datetime,Lat,Lon,TEM")
}
\arguments{
\item{timeRange}{: time range for retrieve, "\link{YYYYMMDDHHMISS,YYYYMMDDHHMISS}",
like, "\link{201509010000,20150903060000}"}

\item{minLat}{: minimum latitude}

\item{minLon}{: minimum longitude}

\item{maxLat}{: maximum latitude}

\item{maxLon}{: maximum longitude}

\item{dataCode}{: dataset code, like "SURF_CHN_MUL_HOR", "SURF_CHN_MUL_HOR_N", and so on.}

\item{elements}{: elements for retrieve, 'ele1,ele2,...'}
}
\value{
data frame of observations.
}
\description{
Retrieve data by cimiss music API.
You should install MUSIC java client.
1) download MUSIC java client from http://10.20.76.55/cimissapiweb/page/apipages/ClientDocumentRight.jsp
2) set environment variable
CLASSPATH=C:/soft/music-lib/Ice.jar;C:/soft/music-lib/music-client-v1.4.0.jar;C:/soft/music-lib
3) call java function
}
\examples{
    elements <- "Station_Id_C,Station_Id_d,Station_Name,Station_levl,Datetime,Lat,Lon,PRE_Time_0808"
    data <- cimiss_obs_in_Rect_by_time_range("[20060801000000,20060801000000]",35,110,45,120,dataCode="SURF_CHN_MUL_DAY",elements=elements)

}
