% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{dt2matrix}
\alias{dt2matrix}
\title{Convert data.table (lon, lat, var1) to matrix (lon, lat)}
\usage{
dt2matrix(data, formula = lon ~ lat, value.var = "var1")
}
\arguments{
\item{data}{: data.table数据, like (lon, lat, var1)}

\item{formula}{: A formula of the form LHS ~ RHS to cast, 给出经纬度的列表名}

\item{value.var}{: 矩阵的数值变量}
}
\value{
: data.table matrix.
}
\description{
Convert data.table (lon, lat, var1) to matrix (lon, lat)
}
