% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_data.R
\name{find_image_coordinates}
\alias{find_image_coordinates}
\title{Translate the given long/lat coordinates into an image position (x, y).}
\usage{
find_image_coordinates(long, lat, bbox, image_width, image_height)
}
\arguments{
\item{long}{longitude value}

\item{lat}{latitude value}

\item{bbox}{bounding box coordinates (list of 2 points with long/lat values)}

\item{image_width}{image width, in pixels}

\item{image_height}{image height, in pixels}
}
\value{
named list with elements "x" and "y" defining an image position
}
\description{
Translate the given long/lat coordinates into an image position (x, y).
}
