% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_data_service.R
\name{gds_concate_url}
\alias{gds_concate_url}
\title{concate gds service url}
\usage{
gds_concate_url(requestType, directory, filename, filter)
}
\arguments{
\item{requestType}{:request type.}

\item{directory}{: the directory on the micaps service}

\item{filter}{: the filename filter pattern, when filename=NULL, this will be used to
find the specified file.}

\item{fileName}{: the data filename, if not given, will use the lastest model run.}
}
\value{
data message.
}
\description{
concate gds service url
}
