% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_data_service.R
\name{gds_get_content}
\alias{gds_get_content}
\title{Retrieve data contents from micaps cassandra service.}
\usage{
gds_get_content(directory, filename)
}
\arguments{
\item{directory}{: the directory on the micaps service}

\item{filename}{: the data filename, if not given, will use the lastest model run.}
}
\value{
raw, bytearray (convert with readBin)
}
\description{
Retrieve data contents from micaps cassandra service.
}
