% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_data_service.R
\name{gds_get_latest_filename}
\alias{gds_get_latest_filename}
\title{Get the latest filename in the data directory on micaps cassandra service.}
\usage{
gds_get_latest_filename(directory, filter = "*.024")
}
\arguments{
\item{directory}{: the directory on the micaps service}

\item{filter}{: the filename filter pattern, when filename=NULL, this will be used to
find the specified file.}
}
\value{
filename, string.
}
\description{
Get the latest filename in the data directory on micaps cassandra service.
}
