% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cache_file}
\alias{get_cache_file}
\title{Get the cache file pathname.}
\usage{
get_cache_file(subDir, filename, name = NULL, cacheClear = TRUE)
}
\arguments{
\item{subDir}{: sub directory string.}

\item{filename}{:  cache filename}

\item{name}{: cache name, like "MICAPS_DATA"}

\item{cacheClear}{: if True, clear old cache folder}
}
\value{

}
\description{
Get the cache file pathname.
}
