% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_from_rcfile}
\alias{get_config_from_rcfile}
\title{Get configure information from config_met_io.ini file.}
\usage{
get_config_from_rcfile()
}
\value{
list, configure parameters.
}
\description{
Get configure information from config_met_io.ini file.
}
