% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{get_http_result}
\alias{get_http_result}
\title{Get CIMISS server data.}
\usage{
get_http_result(interfaceId, params)
}
\arguments{
\item{interfaceId, }{MUSIC interface id.}

\item{params, }{named list for MUSIC parameters.}
}
\value{
json list.
}
\description{
Retrieve the CIMISS data using REST API with pure R code.

refer to:
http://10.20.76.55/cimissapiweb/MethodData_list.action
}
