% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_music_time_range_str}
\alias{get_music_time_range_str}
\title{Get the time range string for MUSIC API.}
\usage{
get_music_time_range_str(startTimeStr, num = 36, units = "hour")
}
\arguments{
\item{startTimeStr}{: start time string, UTC, "YYYYMMDDHH"}

\item{num}{: the time range number.}

\item{units}{: the time units to be used, like "day", "hour", "second", ...}
}
\value{
character.
}
\description{
Get the time range string for MUSIC API.
}
\examples{
  startTimeStr <- "2020021400"
  get_music_time_range_str(startTimeStr)
  
}
