% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cmadaas.R
\name{get_rest_result}
\alias{get_rest_result}
\title{Get the http result from CMADaaS REST api service}
\usage{
get_rest_result(interface_id, params, urlOnly = FALSE)
}
\arguments{
\item{interface_id, }{MUSIC interface id.}

\item{params, }{dictionary for MUSIC parameters.
可以直接指定'serviceNodeId', 'userId'和'pwd'三个参数,
若没有, 则从 .nmcdev 配置文件中读取.}

\item{urlOnly, }{if urlOnly = True, return URL string.}
}
\value{

}
\description{
需要在"/用户名/.nmcdev/config.ini"配置文件中设置CMADaaS数据库地址和账户信息:
\itemize{
\item \link{CMADaaS}
\item DNS = xx.xx.xx.xx
\item PORT = xx
\item USER_ID = test
\item PASSWORD = 123456
\item serviceNodeId = NMIC_MUSIC_CMADAAS
}
}
\examples{
  params <- list('serviceNodeId'='NMIC_MUSIC_CMADAAS', 'userId'='test',
                 'pwd'='123456', 'dataCode'='SURF_CHN_MUL_HOR_N',
                 'elements'='Datetime,Station_Id_d,Lat,Lon,PRE_24h',
                 'times'='20200910000000', 'limitCnt'='10')
  data <- get_rest_result('getSurfEleByTime', params)
  
}
