% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micaps.R
\name{read_micaps_11}
\alias{read_micaps_11}
\title{Read micaps type 11 file (grid).}
\usage{
read_micaps_11(filename, outList = TRUE)
}
\arguments{
\item{filename}{: file name.}

\item{outList}{: if TRUE, return list, or return data.table}
}
\value{
: a list include:
headInfo, head information
dataTime, data date time
fhour, forecast hours
level, data level
lon, grid longitude
lat, grid laitude
dataU, a matrix with dimension c(nlon,nlat)
dataV, a matrix with dimension c(nlon,nlat)
or data.table.
}
\description{
Read micaps type 11 file (grid).
}
\examples{

dataV <- read_micaps_11("Z:\\\\data\\\\newecmwf_grib\\\\pressure\\\\17010108.006")

}
