% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micaps.R
\name{read_micaps_14}
\alias{read_micaps_14}
\title{Read micaps 14 file.}
\usage{
read_micaps_14(fileName)
}
\arguments{
\item{fileName}{: file name.}
}
\value{
A list include draw elements:
lineData,          lines
lineSymData,       line symbols, present trough, front line and so on
symData,           symbols
closedContourData, colosed contour
stationsData,      station situation
weatherRegionData, weather region
fillAreaData,      filled area
notesSymData,      notes symbol
plinesSymData,     line symbols with properties
}
\description{
Read micaps 14 file.
}
\examples{
    fileName <- "Z:/diamond/update/rr010908.024"
    drawElements <- read_micaps_14(fileName)


}
