% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micaps.R
\name{read_micaps_3}
\alias{read_micaps_3}
\title{Read micaps type 3 file (general scatter point).}
\usage{
read_micaps_3(filename)
}
\arguments{
\item{filename}{: file name}
}
\value{
a list include datatime. data.frame (ID,lon,lat,alt,V1,...) and so on.
if read false, return NULL.
}
\description{
Read micaps type 3 file (general scatter point).
}
\examples{
  dataV <- read_micaps_3("Z:\\\\data\\\\surface\\\\jiany_rr\\\\r20\\\\17012108.000")


}
