% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micaps.R
\name{read_micaps_4}
\alias{read_micaps_4}
\title{Read micaps type 4 file (grid).}
\usage{
read_micaps_4(filename)
}
\arguments{
\item{filename}{: file name.}
}
\value{
: a list include:
headInfo, head information
dataTime, data date time
fhour, forecast hours
level, data level
lon, grid longitude
lat, grid laitude
}
\description{
Read micaps type 4 file (grid).
}
\examples{

dataV <- read_micaps_4("Z:\\\\data\\\\newecmwf_grib\\\\pressure\\\\17010108.006")

}
