% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_history.R
\name{retrieve_history_surf_daily_obs}
\alias{retrieve_history_surf_daily_obs}
\title{Download day history observations, one file for each month.}
\usage{
retrieve_history_surf_daily_obs(
  dateRange,
  elements = NULL,
  staLevels = NULL,
  outFile = "day_rain_obs",
  outDir = "."
)
}
\arguments{
\item{dateRange}{: date range, c('YYYYMM','YYYYMM')}

\item{elements}{: elements for retrieve, 'ele1,ele2,...'}

\item{outDir}{: data output directory}
}
\value{
None.
}
\description{
Download day history observations, one file for each month.
}
\examples{
    retrieve_history_surf_daily_obs(c('201001','201512'))


}
