% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_micaps_server.R
\name{retrieve_micaps_model_grid}
\alias{retrieve_micaps_model_grid}
\title{Retrieve micaps grid data from cassandra service.}
\usage{
retrieve_micaps_model_grid(directory, filename = NULL,
  filter = "*.024", outList = FALSE)
}
\arguments{
\item{directory}{: the directory on the micaps service}

\item{filename}{: the data filename, if not given, will use the lastest model run.}

\item{filter}{: the filename filter pattern, when filename=NULL, this will be used to
find the specified file.}

\item{outList}{: if TRUE, return list}
}
\value{
data.table, or list(x, y, z, level, time, fhour)
}
\description{
Retrieve micaps grid data from cassandra service.
}
