% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_micaps_server.R
\name{retrieve_micaps_model_grids}
\alias{retrieve_micaps_model_grids}
\title{Get model forecast time series.}
\usage{
retrieve_micaps_model_grids(
  directory,
  initTimeStr,
  fhours,
  allExists = TRUE,
  cache = TRUE
)
}
\arguments{
\item{directory}{: the directory on the micaps service}

\item{initTimeStr}{: the model forecast initial time string}

\item{fhours}{: the model forecast hours}

\item{allExists}{: all files should exist, or return None.}

\item{cache}{: if TRUE, cache retrieved data to local directory}
}
\value{

}
\description{
Get model forecast time series.
}
\examples{
  initTimeStr <- "20022820"
  fhours <- seq(0, 72, by=6)
  data <- retrieve_micaps_model_grids("ECMWF_HR/TMP/850/", initTimeStr, fhours)
  
}
