% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_micaps_server.R
\name{retrieve_micaps_station_data}
\alias{retrieve_micaps_station_data}
\title{Retrieve micaps station data from cassandra service.}
\usage{
retrieve_micaps_station_data(directory, filename = NULL, filter = "*.000")
}
\arguments{
\item{directory}{: the directory on the micaps service}

\item{filename}{: the data filename, if not given, will use the lastest model run.}

\item{filter}{: the filename filter pattern, when filename=NULL, this will be used to
find the specified file.}
}
\value{
: data.frame of records
}
\description{
Retrieve micaps station data from cassandra service.
}
\examples{
  obs <- retrieve_micaps_station_data("SURFACE/PLOT_NATIONAL/", filename="20190406140000.000")

}
