% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_micaps_server.R
\name{retrieve_micaps_station_data_var_name}
\alias{retrieve_micaps_station_data_var_name}
\title{Map var id to name.}
\usage{
retrieve_micaps_station_data_var_name(vid)
}
\arguments{
\item{vid:}{integer.}
}
\value{
: variable name.
}
\description{
Map var id to name.
}
