% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{smooth2d}
\alias{smooth2d}
\title{Smooth the 2D data.table fields.}
\usage{
smooth2d(x, y, z, theta = 0.03, nx = 128, ny = 128)
}
\arguments{
\item{x}{: 1D vector, longitude.}

\item{y}{: 1D vector, latitude.}

\item{z}{: 1D vector,}

\item{theta}{: the bandwidth or scale parameter, larger for smoother.}

\item{nx}{: Number of grid point in X coordinate.}

\item{ny}{: Number of grid points in Y coordinate.}
}
\value{
data.table, (x, y, z)
}
\description{
该程序调用fields程序库对二维场进行平滑, 首先将离散点插值到nx,ny的网格点上,
如果输入场是等经纬度网格点, 也可看成离散点, 然后插值到(nx,ny)的格点上.之后
调用image.smooth进行平滑, theta控制平滑度, 值越大越平滑. 返回为data.table
数据, 以便于ggplot绘制图像.
}
