% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{uv2angle}
\alias{uv2angle}
\title{convert u, v components to wind angle}
\usage{
uv2angle(u, v)
}
\arguments{
\item{u, }{u components}

\item{v, }{v components}
}
\value{
wind angle
}
\description{
convert u, v components to wind angle
}
