% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{uv2speed}
\alias{uv2speed}
\title{convert u, v components to wind speed.}
\usage{
uv2speed(u, v)
}
\arguments{
\item{u, }{u components}

\item{v, }{v components}
}
\value{
wind speed
}
\description{
convert u, v components to wind speed.
}
