% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{wind2u}
\alias{wind2u}
\title{convert wind angle and speed to u component.}
\usage{
wind2u(angle, speed)
}
\arguments{
\item{angle, }{wind angle or direction}

\item{speed, }{wind speed or velocity}
}
\value{
wind u component
}
\description{
convert wind angle and speed to u component.
}
