% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{wind2v}
\alias{wind2v}
\title{convert wind angle and speed to v component.}
\usage{
wind2v(angle, speed)
}
\arguments{
\item{angle, }{wind angle or direction}

\item{speed, }{wind speed or velocity}
}
\value{
wind v component
}
\description{
convert wind angle and speed to v component.
}
