% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cimiss_server.R
\name{cimiss_get_obs_latest_time}
\alias{cimiss_get_obs_latest_time}
\title{Get the observation latest time.}
\usage{
cimiss_get_obs_latest_time(dataCode = "SURF_CHN_MUL_HOR", latestTime = 6)
}
\arguments{
\item{dataCode}{: dataset code, like "SURF_CHN_MUL_HOR", "SURF_CHN_MUL_HOR_N", and so on.}

\item{latestTime}{: latestTime > 0, like 2 is return the latest time in 2 hours.}
}
\value{
the latest time, like '20200216020000'
}
\description{
\preformatted{Retrieve data latest times by cimiss music REST API.
Refer to http://10.20.76.55/cimissapiweb/index_index.action
}
}
\examples{
  data <- cimiss_get_model_latest_time('SURF_CHN_MUL_HOR_N', latestTime=12)

}
